include video in hero (not advanced)
