+++ blocks/hero-advanced/hero-advanced.js
@@ -1,6 +1,6 @@
-import { isVideo, loadScript } from '../../scripts/scripts.js';
+// import { isVideo, videoButton, loadScript } from '../../scripts/scripts.js';
 // eslint-disable-next-line object-curly-newline
-import { div, a, span, img } from '../../scripts/dom-helpers.js';
+import { a, span } from '../../scripts/dom-helpers.js';
 import { decorateIcons } from '../../scripts/lib-franklin.js';
 import { loadFragment } from '../fragment/fragment.js';
 import { buildHero } from '../hero/hero.js';
@@ -40,25 +40,6 @@ async function buildMediaGallery(mg) {
   });
 }
 
-export function videoButton(container, button, url) {
-  const videoId = url.pathname.split('/').at(-1).trim();
-  const observer = new IntersectionObserver((entries) => {
-    if (entries.some((e) => e.isIntersecting)) {
-      observer.disconnect();
-      loadScript('https://play.vidyard.com/embed/v4.js');
-      const overlay = div({ id: 'overlay' }, div({
-        class: 'vidyard-player-embed', 'data-uuid': videoId, 'dava-v': '4', 'data-type': 'lightbox', 'data-autoplay': '2',
-      }));
-      container.prepend(overlay);
-      button.addEventListener('click', () => {
-        // eslint-disable-next-line no-undef
-        VidyardV4.api.getPlayersByUUID(videoId)[0].showLightbox();
-      });
-    }
-  });
-  observer.observe(container);
-}
-
 export default async function decorate(block) {
   const h1 = block.querySelector('h1');
   const desktop = block.querySelector('div');
@@ -67,17 +48,17 @@ export default async function decorate(block) {
   const mobile = block.querySelector('div');
   h1.parentNode.insertBefore(mobile.querySelector('div:nth-child(2)'), h1.nextSibling);
   mobile.remove();
-  [...block.querySelectorAll('a')].forEach((link) => {
-    const url = new URL(link);
-    if (isVideo(url)) {
-      const container = link.parentElement;
-      container.classList.add('video-column');
-      const videoIcon = div({ class: 'video-icon' }, img({ src: '/images/play_icon.png' }));
-      container.appendChild(videoIcon);
-      videoButton(container, container.querySelector('img'), url);
-      link.remove();
-    }
-  });
+  // [...block.querySelectorAll('a')].forEach((link) => {
+  //   const url = new URL(link);
+  //   if (isVideo(url)) {
+  //     const container = link.parentElement;
+  //     container.classList.add('video-column');
+  //     const videoIcon = div({ class: 'video-icon' }, img({ src: '/images/play_icon.png' }));
+  //     container.appendChild(videoIcon);
+  //     videoButton(container, container.querySelector('img'), url);
+  //     link.remove();
+  //   }
+  // });
   buildHero(block);
   const mg = block.querySelector('a[href*="/media-gallery"]');
   if (mg) {
